requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.9
apparmor==3.0.4
appdirs==1.4.4
application-utility==1.3.2
arrow==1.2.2
asgiref==3.4.1
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
blessed==1.19.1
bs4==0.0.1
btrfsutil==5.17
CacheControl==0.12.6
celery==5.2.3
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
commonmark==0.9.1
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
defusedxml==0.7.1
discord.py==1.7.3
distlib==0.3.4
distro==1.7.0
Django==4.0
django-cors-headers==3.10.1
django-picklefield==3.0.1
django-q==1.3.9
django-simple-history==3.0.0
django-templated-mail==1.1.1
django-throttle-requests==0.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docopt==0.6.2
docutils==0.18.1
drf-writable-nested==0.6.3
evdev==1.5.0
fake-useragent==0.1.11
file-magic==0.4.0
fonttools==4.33.3
h11==0.13.0
html5lib==1.1
idna==3.3
imgkit==1.2.2
importlib-metadata==4.11.3
itypes==1.2.0
jeepney==0.8.0
Jinja2==3.0.3
keyring==23.5.0
keyutils==0.6
kiwisolver==1.4.2
kombu==5.2.3
lensfun==0.3.3
LibAppArmor==3.0.4
libtorrent==2.0.6
lit==13.0.1.dev0
louis==3.21.0
lxml==4.8.0
manjaro-sdk==0.1
MarkupSafe==2.0.1
matplot==0.1.9
matplotlib==3.5.2
more-itertools==8.12.0
msgpack==1.0.3
multidict==6.0.2
netsnmp-python==1.0a1
npyscreen==4.10.5
numpy==1.22.3
oauthlib==3.1.1
ordered-set==4.0.2
outcome==1.1.0
packaging==21.3
pacman-mirrors==4.23.2
parse==1.19.0
pep517==0.12.0
Pillow==9.1.0
pkginfo==1.8.2
ply==3.11
progress==1.6
prompt-toolkit==3.0.26
pycairo==1.21.0
pycparser==2.21
pycups==2.0.1
pycurl==7.45.1
pyee==8.2.2
Pygments==2.12.0
PyGObject==3.42.1
PyJWT==2.3.0
pyloco==0.0.139
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==1.0.2
Pypubsub==4.0.3
PyQt5==5.15.6
PyQt5-sip==12.10.1
pyquery==1.4.3
PySide6==6.3.0
pysmbc==1.0.23
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
rados==2.0.0
rbd==2.0.0
readme-renderer==35.0
redis==3.5.3
reportlab==3.6.9
requests==2.27.0
requests-html==0.10.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
resolvelib==0.5.5
retrying==1.3.3
rfc3986==2.0.0
rgw==2.0.0
rich==12.4.1
SecretStorage==3.3.2
selenium==4.2.0
shiboken6==6.3.0
SimpleWebSocketServer==0.1.1
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
speedtest-cli==2.1.3
sqlparse==0.4.2
systemd-python==234
team==1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
twine==4.0.0
typing==3.7.4.3
typing-extensions==4.2.0
udiskie==2.4.2
uritemplate==4.1.1
urllib3==1.26.7
ushlex==0.99.1
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
websockets==10.3
wsproto==1.1.0
wxPython==4.0.7.post2
yarl==1.7.2
zipp==3.8.0
